Generate bezier slopes of the line in the middle of the leaf

Generate bezier slopes of the line in the middle of the leaf

Usage

gen_middle_line_slopes(leaf_params = gen_leaf_parameters())

Arguments

leaf_params

parameters generated by gen_leaf_parameters()

Value

A dataframe containing the coordinates of the two control points of the bezier curve defining the midvein of the leaf.

Examples

gen_middle_line_slopes()
#> # A tibble: 2 × 2
#>       x      y
#>   <int>  <dbl>
#> 1    -6 -0.882
#> 2   -13  0.500