Manipulate the angles of the leaves with a normal distribution

This function returns a function which itself returns a numerical vector of length of the number of leaves on the branch.

Usage

spark_norm(mean = 0, sd = 3)

Arguments

mean, sd

Parameters passed to stats::dnorm().

Value

dnorm() function with n_leaves as one of the arguments

Examples

spark_norm()
#> function (n_leaves) 
#> {
#>     stats::rnorm(n_leaves, mean = mean, sd = sd)
#> }
#> <bytecode: 0x55f66f3efb30>
#> <environment: 0x55f678ca19a8>